+++ tests/spec/airconsole-1.11.0-spec.js
@@ -25,7 +25,7 @@ describe("AirConsole 1.11.0", function () {
 
     function tearDown() {
         if (airconsole) {
-            window.removeEventListener('message', airconsole.onPostMessage_);
+            window.removeEventListener('message', airconsole.messageEventListener_);
             airconsole = null;
         }
     }
@@ -383,5 +383,18 @@ describe("AirConsole 1.11.0", function () {
         });
 
         testAirConsole110Plus();
+    });
+
+    /**
+     ======================================================================================
+     TEST MEDIA PERMISSIONS FUNCTIONALITY
+     */
+
+    describe("User Media Permissions", function () {
+        afterEach(function () {
+            tearDown();
+        });
+
+        testUserMediaPermissions();
     });
 });